requirements.txt
python-telegram-bot>=20.0
nest-asyncio
spacy
pymongo
openai
langchain-text-splitters
nltk
langchain-chroma
langchain-openai
langchain_community tiktoken langchainhub chromadb langchain